+++ Reynj.Text.Json/InternalRangeConverter.cs
@@ -25,68 +25,45 @@ namespace Reynj.Text.Json
                 throw new JsonException();
             }
 
-            T start = default!;
+            var valueType = typeof(T);
+
+            T start = default;
             var startSet = false;
 
-            T end = default!;
+            T end = default;
             var endSet = false;
 
-            // Get the first property.
             reader.Read();
-            if (reader.TokenType != JsonTokenType.PropertyName)
+
+            while (reader.TokenType == JsonTokenType.PropertyName)
             {
-                throw new JsonException();
+                var propertyName = reader.GetString()!;
+                if (string.Equals(propertyName, StartName, StringComparison.OrdinalIgnoreCase))
+                {
+                    start = ReadProperty<T>(ref reader, valueType, options);
+                    startSet = true;
+                }
+                else if (string.Equals(propertyName, EndName, StringComparison.OrdinalIgnoreCase))
+                {
+                    end = ReadProperty<T>(ref reader, valueType, options);
+                    endSet = true;
+                }
+                else
+                {
+                    reader.Skip();
+                }
+
+                reader.Read();
             }
 
-            var propertyName = reader.GetString()!;
-            if (propertyName == StartName)
-            {
-                start = ReadProperty<T>(ref reader, typeToConvert, options);
-                startSet = true;
-            }
-            else if (propertyName == EndName)
-            {
-                end = ReadProperty<T>(ref reader, typeToConvert, options);
-                endSet = true;
-            }
-            else
-            {
-                throw new JsonException();
-            }
-
-            // Get the second property.
-            reader.Read();
-            if (reader.TokenType != JsonTokenType.PropertyName)
-            {
-                throw new JsonException();
-            }
-
-            propertyName = reader.GetString()!;
-            if (propertyName == EndName)
-            {
-                end = ReadProperty<T>(ref reader, typeToConvert, options);
-                endSet = true;
-            }
-            else if (propertyName == StartName)
-            {
-                start = ReadProperty<T>(ref reader, typeToConvert, options);
-                startSet = true;
-            }
-            else
+            if (reader.TokenType != JsonTokenType.EndObject)
             {
                 throw new JsonException();
             }
 
             if (!startSet || !endSet)
             {
-                throw new JsonException();
-            }
-
-            reader.Read();
-
-            if (reader.TokenType != JsonTokenType.EndObject)
-            {
-                throw new JsonException();
+                return Range<T>.Empty;
             }
 
             return new Range<T>(start, end);
@@ -96,33 +73,30 @@ namespace Reynj.Text.Json
         public override void Write(Utf8JsonWriter writer, Range<T> value, JsonSerializerOptions options)
         {
             writer.WriteStartObject();
-            WriteProperty(writer, value.Start, _startName, options);
-            WriteProperty(writer, value.End, _endName, options);
+
+            if (!value.IsEmpty())
+            {
+                WriteProperty(writer, value.Start, _startName, options);
+                WriteProperty(writer, value.End, _endName, options);
+            }
+
             writer.WriteEndObject();
         }
 
         private static TValue ReadProperty<TValue>(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-            where TValue : IComparable
         {
-            TValue value;
-
             // Attempt to use existing converter first before re-entering through JsonSerializer.Deserialize().
             // The default converter for objects does not parse null objects as null, so it is not used here.
-            if (typeToConvert != typeof(object) && options?.GetConverter(typeToConvert) is JsonConverter<TValue> keyConverter)
+            if (typeToConvert != typeof(object) && options?.GetConverter(typeToConvert) is JsonConverter<TValue> valueConverter)
             {
                 reader.Read();
-                value = keyConverter.Read(ref reader, typeToConvert, options);
-            }
-            else
-            {
-                value = JsonSerializer.Deserialize<TValue>(ref reader, options);
+                return valueConverter.Read(ref reader, typeToConvert, options);
             }
 
-            return value;
+            return JsonSerializer.Deserialize<TValue>(ref reader, options);
         }
 
         private static void WriteProperty<TValue>(Utf8JsonWriter writer, TValue value, JsonEncodedText name, JsonSerializerOptions options)
-            where TValue : IComparable
         {
             var typeToConvert = typeof(TValue);
 
@@ -130,9 +104,9 @@ namespace Reynj.Text.Json
 
             // Attempt to use existing converter first before re-entering through JsonSerializer.Serialize().
             // The default converter for object does not support writing.
-            if (typeToConvert != typeof(object) && options?.GetConverter(typeToConvert) is JsonConverter<TValue> keyConverter)
+            if (typeToConvert != typeof(object) && options?.GetConverter(typeToConvert) is JsonConverter<TValue> valueConverter)
             {
-                keyConverter.Write(writer, value, options);
+                valueConverter.Write(writer, value, options);
             }
             else
             {